srcs/gpt-assistant/requirements.txt
openai
pathlib
psycopg2-binary